% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacityPlot.homoVAR.fn.R
\name{capacityPlot.homoVAR.fn}
\alias{capacityPlot.homoVAR.fn}
\title{Display assay capacity}
\usage{
capacityPlot.homoVAR.fn(
  FoldHit,
  color = c("green", "black", "red", "purple"),
  pch = 1:4,
  xRange = NULL,
  yRange = NULL,
  normCut = 3,
  nPointSim = 400,
  length.out = 1000,
  round.point = 2,
  lwd = 3,
  mar = c(8.1, 4.1, 1.1, 1.1),
  lineScale = c(0.9, 3.7, 6.4),
  lineXaxis = c(0, 2.8, 5.5)
)
}
\arguments{
\item{FoldHit}{a vector of FoldHit values for controls, the first element must be 0, the second element must be 1 or -1, the remainings are for positive controls}

\item{color}{colors mark negative reference, MDhit and positive controls}

\item{pch}{point types mark negative reference, MDhit and positive controls}

\item{xRange}{range for x-axis}

\item{yRange}{range for y-axis}

\item{normCut}{SSMD value for defining a maximally detectable hit}

\item{nPointSim}{number of points to be simulated for generating a density}

\item{length.out}{number of points for showing density}

\item{round.point}{number of round points for showing capacity values}

\item{lwd}{width of density lines}

\item{mar}{margin for the plot}

\item{lineScale}{value of line in showing scale}

\item{lineXaxis:}{value of line in showing labels in x-axis}
}
\value{
no return value
}
\description{
a function to plot density and simulated values of measured response to demonstrate FoldHit under homoscedasticity for two groups excluding NA values
}
\examples{
  FoldHit = c(0, -1, -4.308057, -1.674495)
  par(mfrow=c(1,2))
  capacityPlot.homoVAR.fn( FoldHit, color=c("green","black","red", "purple"))
  capacityPlot.homoVAR.fn( -FoldHit, color=c("green","black","red", "purple"))

}
\author{
Xiaohua Douglas Zhang  06/2020
}
