% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIempirical.fn.R
\name{CIempirical.fn}
\alias{CIempirical.fn}
\title{Calculate empirical confidence interval}
\usage{
CIempirical.fn(x, alpha = 0.05)
}
\arguments{
\item{x}{a vector of values}

\item{alpha}{to specify 1-alpha confidence interval}
}
\value{
CI
}
\description{
a function to calculate empirical confidence interval
}
\examples{
x = rnorm(1000)
  CIempirical.fn(x, alpha=0.5)
}
\author{
Xiaohua Douglas Zhang  05/2020
}
