% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIt.fn.R
\name{CIt.fn}
\alias{CIt.fn}
\title{calculate confidence interval from t-distribution}
\usage{
CIt.fn(x, alpha = 0.05)
}
\arguments{
\item{x}{a vector of values}

\item{alpha}{to specify 1-alpha confidence interval}
}
\value{
CI
}
\description{
a function to calculate confidence interval of values from the t-distribution
}
\examples{
x = rnorm(1000)
  CIt.fn(x, alpha=0.5)
}
\author{
Xiaohua Douglas Zhang  05/2020
}
