% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldHit.homoVAR.UMVUE.fn.R
\name{FoldHit.homoVAR.UMVUE.fn}
\alias{FoldHit.homoVAR.UMVUE.fn}
\title{Calculate FoldHit}
\usage{
FoldHit.homoVAR.UMVUE.fn(Y1, Y2)
}
\arguments{
\item{Y1}{a value or a vector for the first group (reference group)}

\item{Y2}{a vector for the second group}
}
\value{
FoldHit
}
\description{
a function to calculate the UMVUE of FoldHit under homoscedasticity for two groups excluding NA values
}
\examples{
y1 = rnorm(12, 0, 1); y2 = rnorm(12, 3, 1)
  FoldHit.homoVAR.UMVUE.fn(y1, y2)
}
\author{
: Xiaohua Douglas Zhang        02/2020
}
