% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldHit0.homoVAR.fn.R
\name{FoldHit0.homoVAR.fn}
\alias{FoldHit0.homoVAR.fn}
\title{calculate FoldHit with error control}
\usage{
FoldHit0.homoVAR.fn(Y1, Y2, Alpha = 0.05, length.out = 1000, minDIFF = 0.01)
}
\arguments{
\item{Y1}{a value or a vector for the first group (reference group)}

\item{Y2}{a vector for the second group}

\item{Alpha}{the significant level}

\item{length.out}{the length of vectors of FoldHit0 candidate values for search FoldHit0.}

\item{minDIFF}{If the difference between the input FoldHitC and the calculated FoldHitC corresponding to a FoldHit0 value, then stop the calculation process and return this calculated value as what we want}
}
\value{
FoldHit0
}
\description{
a function to search for FoldHit0 corresponding to the estimated value of FoldHit
}
\examples{
 y1 = rnorm(12, 0, 1); y2 = rnorm(12, -3, 1)
  FoldHit.homoVAR.UMVUE.fn(y1, y2)
  FoldHit0.homoVAR.fn(y1, y2, Alpha=0.05, length.out=1000, minDIFF=0.001)
  y1 = -y1; y2 = -y2
  FoldHit.homoVAR.UMVUE.fn(y1, y2)
  FoldHit0.homoVAR.fn(y1, y2, Alpha=0.05, length.out=1000, minDIFF=0.001)
}
\author{
Xiaohua Douglas Zhang        02/2020
}
