% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldHit0core.homoVAR.fn.R
\name{FoldHit0core.homoVAR.fn}
\alias{FoldHit0core.homoVAR.fn}
\title{Calculate FoldHit with error control in the middle}
\usage{
FoldHit0core.homoVAR.fn(
  FoldHitC,
  n1,
  n2,
  Alpha,
  length.out = 1000,
  minDIFF = 0.01
)
}
\arguments{
\item{FoldHitC}{the critical value of FoldHit}

\item{n1}{sample size for the first group (reference group)}

\item{n2}{sample size for the second group}

\item{Alpha}{significant level}

\item{length.out}{the length of vectors of FoldHit0 candidate values for search FoldHit0.}

\item{minDIFF}{if the difference between the input FoldHitC and the calculated FoldHitC corresponding to a FoldHit0 value, then stop the calculation process and return this calculated value as what we want}
}
\value{
FoldHit0
}
\description{
a core function to search for FoldHit0 corresponding to FoldHitC (FoldHit critical value)
}
\examples{
FoldHit0core.homoVAR.fn(FoldHitC=-1.121588, n1=304, n2=16, Alpha=0.05, minDIFF=0.001)
}
\author{
Xiaohua Douglas Zhang        02/2020
}
