% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldHitC.homoVAR.fn.R
\name{FoldHitC.homoVAR.fn}
\alias{FoldHitC.homoVAR.fn}
\title{Calculate the critical value of FoldHit}
\usage{
FoldHitC.homoVAR.fn(n1, n2, Alpha, FoldHit0 = 1)
}
\arguments{
\item{n1}{sample size for the first group (reference group)}

\item{n2}{sample size for the second group}

\item{Alpha}{significant level}

\item{FoldHit0}{population value of FoldHit}
}
\value{
FoldHitC
}
\description{
a function to calculate the critical value of FoldHit (i.e., FoldHitC)
}
\examples{
FoldHitC.homoVAR.fn(304, 16, Alpha=0.05, FoldHit0=1)
}
\author{
Xiaohua Douglas Zhang        02/2020
}
