% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateWelltoX.fn.R
\name{plateWelltoX.fn}
\alias{plateWelltoX.fn}
\title{Convert plate, row and column numbers to a series of values}
\usage{
plateWelltoX.fn(data.df, nRow = 16, nCol = 24, byRow = TRUE)
}
\arguments{
\item{data.df}{consists of three columns: plate names or index, rows and columns}

\item{nRow}{number of rows in a plate}

\item{nCol}{number of columns in a plate}

\item{byRow}{whether the well series is arranged by row}
}
\value{
a data frame
}
\description{
a function to generate value in x-axis based on plate number and well positions in a plate.
}
\examples{
plateID = rep(1:5, each=384)
  rowCore = rep(1:16, each=24); rows = rep(rowCore, 5)
  columnCore = rep(1:24, 16); columns = rep(columnCore, 5)
  data.df = data.frame(plateID, rows, columns)
  x.df = plateWelltoX.fn(data.df, nRow=16, nCol=24, byRow=TRUE)

}
\author{
Xiaohua Douglas Zhang   2017
}
