% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zFactor.fn.R
\name{zFactor.fn}
\alias{zFactor.fn}
\title{Calculate z factor}
\usage{
zFactor.fn(x, y, k = 3)
}
\arguments{
\item{x}{a value or a vector for the first group (reference group)}

\item{y}{a value of a vector for the second group}
}
\description{
a function to calculate plug-in estimate of z factor
}
\examples{
x=rnorm(20)
  y=rnorm(30)
  zFactor.fn(x, y)
}
\author{
Xiaohua Douglas Zhang
}
